# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: DynamicsERP API specification — Solar Projects ERP
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /vendors:
    get:
      operationId: listVendors
      tags: [vendors]
      parameters:
        - { in: query, name: search, schema: { type: string } }
        - { in: query, name: isActive, schema: { type: boolean } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { type: array, items: { $ref: "#/components/schemas/Vendor" } } } } }
    post:
      operationId: createVendor
      tags: [vendors]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/CreateVendor" } } } }
      responses:
        "201": { description: Created, content: { application/json: { schema: { $ref: "#/components/schemas/Vendor" } } } }
  /vendors/{id}:
    get:
      operationId: getVendor
      tags: [vendors]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Vendor" } } } }
    patch:
      operationId: updateVendor
      tags: [vendors]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/UpdateVendor" } } } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Vendor" } } } }
    delete:
      operationId: deleteVendor
      tags: [vendors]
      parameters: [{ in: path, name: id, required: true, schema: { type: integer } }]
      responses:
        "204": { description: Deleted }
...
components:
  schemas:
...
    Vendor:
      type: object
      properties:
        id: { type: integer }
        code: { type: string }
        name: { type: string }
        gstin: { type: ["string", "null"] }
        pan: { type: ["string", "null"] }
        contactPerson: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        addressLine: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        pincode: { type: ["string", "null"] }
        bankName: { type: ["string", "null"] }
        bankAccountNo: { type: ["string", "null"] }
        ifscCode: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        rating: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required: [id, code, name, isActive, createdAt, updatedAt]
    CreateVendor:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        gstin: { type: ["string", "null"] }
        pan: { type: ["string", "null"] }
        contactPerson: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        addressLine: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        pincode: { type: ["string", "null"] }
        bankName: { type: ["string", "null"] }
        bankAccountNo: { type: ["string", "null"] }
        ifscCode: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        rating: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
      required: [code, name]
    UpdateVendor:
      type: object
      properties:
        name: { type: ["string", "null"] }
        gstin: { type: ["string", "null"] }
        pan: { type: ["string", "null"] }
        contactPerson: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        addressLine: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        state: { type: ["string", "null"] }
        pincode: { type: ["string", "null"] }
        bankName: { type: ["string", "null"] }
        bankAccountNo: { type: ["string", "null"] }
        ifscCode: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        rating: { type: ["integer", "null"] }
        notes: { type: ["string", "null"] }
        isActive: { type: ["boolean", "null"] }
...